refactor: move validate tests to new file
validate.rs was too big
